fix(backend): 絵文字の変換処理が不十分なのを修正
